GameStatus tests done

+++ 393 Demo 1/Assets/Scripts/UIScripts/LeaderboardScript.cs
@@ -3,6 +3,7 @@ using UnityEngine.UI;
 
 public class LeaderboardScript : MonoBehaviour
 {
+    public static UIManager uim = new UIManager("Highscores");
     public Text highScore1;
     public Text highScore2;
     public Text highScore3;
@@ -16,11 +17,14 @@ public class LeaderboardScript : MonoBehaviour
 
     public void Reset ()
     {
-        PlayerPrefs.DeleteKey("FirstPlace");
-        PlayerPrefs.DeleteKey("SecondPlace");
-        PlayerPrefs.DeleteKey("ThirdPlace");
-        highScore1.text = "1st Place    0";
-        highScore2.text = "2nd Place    0";
-        highScore3.text = "3rd Place    0";
+        if(uim.ClickReset())
+        {
+            PlayerPrefs.DeleteKey("FirstPlace");
+            PlayerPrefs.DeleteKey("SecondPlace");
+            PlayerPrefs.DeleteKey("ThirdPlace");
+            highScore1.text = "1st Place    0";
+            highScore2.text = "2nd Place    0";
+            highScore3.text = "3rd Place    0";
+        }
     }
 }
